Add files via upload
An archaic batch file I made way back in the day. If you find an ip address to direct it to it will start a bunch of processes, very crudely, and start pinging the ip; Acting as a DOS attack. I would use this on our schools routers that for some reason would take and process every ping I gave it. I tried it on my old router, and literally destroyed it. We were out of a router for 5 days. WILL MAKE YOUR COMPUTER HAVE A SEIZURE.